apply simf to metrics
