+++ lib/segment/src/spaces/metric_uint/avx2/euclid.rs
