+++ lib/segment/src/spaces/metric_uint/avx2/manhattan.rs
