+++ lib/segment/src/spaces/metric_uint/neon/dot.rs
