+++ lib/segment/src/spaces/metric_uint/neon/euclid.rs
