+++ lib/segment/src/spaces/metric_uint/simple_euclid.rs
@@ -3,11 +3,11 @@ use common::types::ScoreType;
 use crate::data_types::vectors::{DenseVector, TypedDenseVector, VectorElementTypeByte};
 use crate::spaces::metric::Metric;
 #[cfg(target_arch = "x86_64")]
-use crate::spaces::metric_uint::avx2::avx_euclid::avx_euclid_similarity_bytes;
+use crate::spaces::metric_uint::avx2::euclid::avx_euclid_similarity_bytes;
 #[cfg(all(target_arch = "aarch64", target_feature = "neon"))]
-use crate::spaces::metric_uint::neon::neon_euclid::neon_euclid_similarity_bytes;
+use crate::spaces::metric_uint::neon::euclid::neon_euclid_similarity_bytes;
 #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
-use crate::spaces::metric_uint::sse2::sse_euclid::sse_euclid_similarity_bytes;
+use crate::spaces::metric_uint::sse2::euclid::sse_euclid_similarity_bytes;
 #[cfg(target_arch = "x86_64")]
 use crate::spaces::simple::MIN_DIM_SIZE_AVX;
 use crate::spaces::simple::{EuclidMetric, MIN_DIM_SIZE_SIMD};
