+++ lib/segment/src/spaces/metric_uint/sse2/dot.rs
