+++ lib/segment/src/spaces/metric_uint/sse2/euclid.rs
