+++ lib/segment/src/spaces/metric_uint/sse2/manhattan.rs
